vehicle-ai-concierge/requirements.txt
flask
python-dotenv
google-genai
google-adk
google-cloud-storage
google-cloud-aiplatform
google-cloud-discoveryengine==0.11.14
gunicorn